+++ src/sentry/notifications/platform/templates/workflow_engine/__init__.py
@@ -1,0 +1,1 @@
+from .activity import *  # noqa: F401,F403
