+++ src/sentry/notifications/platform/templates/workflow_engine/activity/__init__.py
@@ -1,0 +1,20 @@
+from .seer_base import ACTIVITY_TYPE_TO_SOURCE, WorkflowEngineActivityAction
+from .seer_coding_completed import SeerCodingCompletedActivityTemplate
+from .seer_coding_started import SeerCodingStartedActivityTemplate
+from .seer_pr_created import SeerPrCreatedActivityTemplate
+from .seer_rca_completed import SeerRcaCompletedActivityTemplate
+from .seer_rca_started import SeerRcaStartedActivityTemplate
+from .seer_solution_completed import SeerSolutionCompletedActivityTemplate
+from .seer_solution_started import SeerSolutionStartedActivityTemplate
+
+__all__ = (
+    "ACTIVITY_TYPE_TO_SOURCE",
+    "WorkflowEngineActivityAction",
+    "SeerRcaStartedActivityTemplate",
+    "SeerRcaCompletedActivityTemplate",
+    "SeerSolutionStartedActivityTemplate",
+    "SeerSolutionCompletedActivityTemplate",
+    "SeerCodingStartedActivityTemplate",
+    "SeerCodingCompletedActivityTemplate",
+    "SeerPrCreatedActivityTemplate",
+)
